[webtranport] Update streamErrorCode to 32-bit long
The latest version of WebTransport protocol uses 32-bit unsigned
integer for streamErrorCode [1, 2, 3].

[1] https://github.com/ietf-wg-webtrans/draft-ietf-webtrans-http3/pull/115
[2] https://github.com/w3c/webtransport/pull/509
[3] https://github.com/w3c/webtransport/pull/527

Bug: 1465282
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4693162
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1172268}

+++ services/network/web_transport_unittest.cc
@@ -223,8 +223,8 @@ class TestClient final : public mojom::WebTransportClient {
       std::move(quit_closure_for_outgoing_stream_closure_).Run();
     }
   }
-  void OnReceivedResetStream(uint32_t stream_id, uint8_t) override {}
-  void OnReceivedStopSending(uint32_t stream_id, uint8_t) override {}
+  void OnReceivedResetStream(uint32_t stream_id, uint32_t) override {}
+  void OnReceivedStopSending(uint32_t stream_id, uint32_t) override {}
   void OnClosed(mojom::WebTransportCloseInfoPtr close_info) override {}
 
   void WaitUntilMojoConnectionError() {
